--- react/components/Table/util/actionTypes.js
@@ -1,13 +1,0 @@
-export default {
-  SET_DENSITY: 0,
-  SET_SELECTED_ROWS: 1,
-  SET_ALL_LINES_SELECTED: 2,
-  SET_HIDDEN_FIELDS: 3,
-  SET_TABLE_ROW_HEIGHT: 4,
-  DESELECT_ALL_ROWS: 5,
-  SELECT_ALL_ROWS: 6,
-  SELECT_ROW: 7,
-  HIDE_COLUMN: 8,
-  HIDE_ALL_COLUMNS: 9,
-  UNHIDE_ALL_COLUMNS: 10,
-}
