--- react/components/Table/util/constants.js
@@ -1,12 +1,0 @@
-export default {
-  TABLE_HEADER_HEIGHT: 36,
-  EMPTY_STATE_SIZE_IN_ROWS: 5,
-  DEFAULT_SCROLLBAR_WIDTH: 17,
-  DENSITY_OPTIONS: ['low', 'medium', 'high'],
-  FIELDS_BOX_ITEM_HEIGHT: 36,
-  BOX_SHADOW_STYLE: { boxShadow: '0px 1px 18px rgba(0, 0, 0, 0.14)' },
-  MEDIUM_ICON_SIZE: 14,
-  MAX_FIELDS_BOX_HEIGHT: 192,
-  FIELDS_BOX_WIDTH: 292,
-  LIGHT_ICON_SIZE: 16,
-}
