--- react/components/Table/util/getInitialHiddenFieldsFromSchema.js
@@ -1,9 +1,0 @@
-export default schema => {
-  const hiddenFields = []
-  Object.keys(schema.properties).forEach(key => {
-    if (schema.properties[key].hidden) {
-      hiddenFields.push(key)
-    }
-  })
-  return hiddenFields
-}
