--- react/components/Table/util/index.js
@@ -1,17 +1,0 @@
-import actionTypes from './actionTypes'
-import calculateTableHeight from './calculateTableHeight'
-import constants from './constants'
-import getInitialHiddenFieldsFromSchema from './getInitialHiddenFieldsFromSchema'
-import getRowHeight from './getRowHeight'
-import getScrollbarWidth from './getScrollbarWidth'
-import reducer from './reducer'
-
-export {
-  actionTypes,
-  calculateTableHeight,
-  constants,
-  getInitialHiddenFieldsFromSchema,
-  getRowHeight,
-  getScrollbarWidth,
-  reducer,
-}
